итерация 2
