+++ ewm-service/src/main/java/ru/practicum/controller/admin/EventControllerAdmin.java
@@ -1,0 +1,54 @@
+package ru.practicum.controller.admin;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.dto.event.EventFullDto;
+import ru.practicum.dto.input.UpdateEventAdminRequest;
+import ru.practicum.service.EventService;
+
+import javax.validation.Valid;
+import javax.validation.constraints.Min;
+import java.time.LocalDateTime;
+import java.util.List;
+
+/**
+ * @author MR.k0F31N
+ */
+
+@RestController
+@Slf4j
+@RequiredArgsConstructor
+@Validated
+@RequestMapping("/admin/events")
+public class EventControllerAdmin {
+    private final EventService eventService;
+
+    @GetMapping
+    public List<EventFullDto> searchEvents(@RequestParam(required = false) List<Long> users,
+                                           @RequestParam(required = false) List<String> states,
+                                           @RequestParam(required = false) List<Long> categories,
+                                           @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
+                                           @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeEnd,
+                                           @RequestParam(value = "from", required = false, defaultValue = "0") @Min(0) Integer from,
+                                           @RequestParam(value = "size", required = false, defaultValue = "10") @Min(1) Integer size) {
+        log.trace("Endpoint request: GET /admin/events");
+        log.debug("Param: search id user = '{}', search state events = '{}', search id categories = '{}', range start = '{}', " +
+                "range end = '{}', pageable from = '{}', pageable size = '{}'", users, states, categories, rangeStart, rangeEnd, from, size);
+        final Pageable pageable = PageRequest.of(from / size, size, Sort.by(Sort.Direction.ASC, "id"));
+        return eventService.getAllEventForParamFromAdmin(users, states, categories, rangeStart, rangeEnd, pageable);
+    }
+
+    @PatchMapping("/{eventId}")
+    public EventFullDto updateEventByAdmin(@PathVariable(value = "eventId") @Min(1) Long eventId,
+                                           @RequestBody @Valid UpdateEventAdminRequest inputUpdate) {
+        log.trace("Endpoint request: PATCH /admin/events/{eventId}");
+        log.debug("Param: event id = '{}', object update = '{}'", eventId, inputUpdate);
+        return eventService.updateEventByEventIdFromAdmin(eventId, inputUpdate);
+    }
+}
